Implement plan: extract AI.ResponseProcessor from AI context
